Updated battery sensor config.
