+++ Autopilot/Path Manager/BatterySensor.h
