+++ Autopilot/Path Manager/main.h
@@ -14,7 +14,7 @@
 
 //Turns debug mode on or off. Usually contains small snippets of code to output
 //data through UART or to provide small input adjustments
-#define DEBUG 0
+#define DEBUG 1
 
 //  Path Manager - Communicates with the GPS in order to provide a constant
 //                  heading for the aircraft to follow.
